Font now Integrated into the application
This is needs testing on your computers too.

The issue was we were following the correct process to add the font into
the application resources, but we needed to amend one of the setting to
make that file an embed resource...

Also it don't seem to like having two fonts added.