Tidy Up
Tidying up the code to put it in a more logical order, almost in order
as the program runs. But now makes more sense.

Also added back in the highlighting of the buttons, this still glitches
on the exit button. Hopefully this gets fixed before were done, but it's
not essential for the running of the game.